requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai